refactor: move TestCacheHome into tombi-test-lib
