+++ crates/tombi-cache/Cargo.toml
@@ -16,3 +16,4 @@ tombi-uri.workspace = true
 [dev-dependencies]
 tempfile.workspace = true
 tokio = { workspace = true, features = ["macros", "rt"] }
+tombi-test-lib.workspace = true
