+++ crates/tombi-extension/Cargo.toml
@@ -25,3 +25,4 @@ tower-lsp.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = ["macros", "rt"] }
 tombi-schema-store = { workspace = true, features = ["native"] }
+tombi-test-lib.workspace = true
