+++ crates/tombi-test-lib/Cargo.toml
@@ -9,4 +9,5 @@ license.workspace = true
 [dependencies]
 chrono.workspace = true
 env_logger.workspace = true
+tempfile.workspace = true
 tombi-uri.workspace = true
