+++ crates/tombi-test-lib/src/lib.rs
@@ -1,5 +1,7 @@
+mod cache;
 mod date_time;
 mod path;
+pub use cache::*;
 pub use date_time::*;
 pub use path::*;
