+++ extensions/tombi-extension-cargo/Cargo.toml
@@ -33,3 +33,4 @@ tower-lsp.workspace = true
 [dev-dependencies]
 tempfile.workspace = true
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
+tombi-test-lib.workspace = true
